superpixel clustering polished

+++ .src/image_io.inl
@@ -23,9 +23,9 @@ namespace crisp
         ColorImage out;
         out.create(image.getSize().x, image.getSize().y);
 
-        for (long x = 0; x < image.getSize().x; ++x)
+        for (size_t x = 0; x < image.getSize().x; ++x)
         {
-            for (long y = 0; y < image.getSize().y; ++y)
+            for (size_t y = 0; y < image.getSize().y; ++y)
             {
                 auto color = image.getPixel(x, y);
                 out(x, y) = RGB(color.r / 255.f, color.g / 255.f, color.b / 255.f);
@@ -47,9 +47,9 @@ namespace crisp
         GrayScaleImage out;
         out.create(image.getSize().x, image.getSize().y);
 
-        for (long x = 0; x < image.getSize().x; ++x)
+        for (size_t x = 0; x < image.getSize().x; ++x)
         {
-            for (long y = 0; y < image.getSize().y; ++y)
+            for (size_t y = 0; y < image.getSize().y; ++y)
             {
                 auto color = image.getPixel(x, y);
                 out(x, y) = (color.r / 255.f + color.g / 255.f + color.b / 255.f) / 3.f;
@@ -73,9 +73,9 @@ namespace crisp
 
         bool had_to_threshold = false;
 
-        for (long x = 0; x < image.getSize().x; ++x)
+        for (size_t x = 0; x < image.getSize().x; ++x)
         {
-            for (long y = 0; y < image.getSize().y; ++y)
+            for (size_t y = 0; y < image.getSize().y; ++y)
             {
                 auto color = image.getPixel(x, y);
 
@@ -102,9 +102,9 @@ namespace crisp
         sf::Image image;
         image.create(in.get_size().x(), in.get_size().y());
 
-        for (long x = 0; x < in.get_size().x(); ++x)
+        for (size_t x = 0; x < in.get_size().x(); ++x)
         {
-            for (long y = 0; y < in.get_size().y(); ++y)
+            for (size_t y = 0; y < in.get_size().y(); ++y)
             {
                 RGB color_in = in(x, y);
                 image.setPixel(x, y, sf::Color(uint8_t(color_in.red() * 255),
@@ -128,9 +128,9 @@ namespace crisp
         sf::Image image;
         image.create(in.get_size().x(), in.get_size().y());
 
-        for (long x = 0; x < in.get_size().x(); ++x)
+        for (size_t x = 0; x < in.get_size().x(); ++x)
         {
-            for (long y = 0; y < in.get_size().y(); ++y)
+            for (size_t y = 0; y < in.get_size().y(); ++y)
             {
                 float intensity = in(x, y);
                 image.setPixel(x, y, sf::Color(uint8_t(intensity * 255),
@@ -154,9 +154,9 @@ namespace crisp
         sf::Image image;
         image.create(in.get_size().x(), in.get_size().y());
 
-        for (long x = 0; x < in.get_size().x(); ++x)
+        for (size_t x = 0; x < in.get_size().x(); ++x)
         {
-            for (long y = 0; y < in.get_size().y(); ++y)
+            for (size_t y = 0; y < in.get_size().y(); ++y)
             {
                 bool which = in(x, y);
